Big ActiveActor refactoring
1. Refactor ActiveActor not to subclass ImageView, allowing for more complex
   actors in the future. One example is Boss, who is now able to
   encapsulate the Shield (will be done in a future commit).
2. Introduce ActorFactory.
3. Introduce AssetFactory interface and CachedAssetFactory as it's
   implementation. This will stop images having to be loaded every time
   a new actor/ui element is created, potentially reducing memory usage.
4. Migrate all deltaTime calculations to use a double value in seconds.
   This will allow for more precision and easier calculations.

+++ src/main/java/com/example/demo/screen/AbstractScreen.java
@@ -90,7 +90,7 @@ public abstract class AbstractScreen {
      * Subclasses must override this method to define screen-specific behavior during
      * each update cycle.
      *
-     * @param timeDelta the time elapsed since the last update, in milliseconds.
+     * @param timeDelta the time elapsed since the last update, in seconds.
      */
-    public abstract void updateScene(int timeDelta);
+    public abstract void updateScene(double timeDelta);
 }
